requirements.txt
behave==1.2.6
pika==0.13.0
PyHamcrest==1.9.0
requests==2.21.0
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
Flask-Login==0.4.1
pymongo==3.8.0